render photos from portfolio
